ref: Remove symbolication/symbolication.rs
